requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database & Migrations
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.25

# Authentication & Security
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
email-validator==2.1.0.post1

# Forms & Validation
Flask-WTF==1.2.1
WTForms==3.1.2

# Background Tasks (Celery)
Celery==5.3.6
redis==5.0.1

# PDF Processing (The Engine)
PyPDF2==3.0.1          # Basic splitting/merging
pikepdf==8.11.2        # Advanced PDF repair/manipulation (Enterprise grade)
reportlab==4.0.9       # Generating new PDFs
pdf2image==1.17.0      # Rendering PDF to images for preview

# Utilities
python-dotenv==1.0.0   # Managing environment variables
gunicorn==21.2.0       # Production server
marshmallow==3.20.2    # Data serialization (for API)